+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [2023.2.11]
 
 ### `Project Import` enhancements
+- Introduced customization of the XSD schema for `y` namespace [#786](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/786)
 - Improved classpath configuration for OOTB modules and readonly mode [#783](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/783)
 
 ### `Cockpit NG` enhancements